Do not ignore an old write when it comes after a newer write
Igoring an old write causes a weird problem in the following scenario:

1) Txn executes a put operation with time T.
2) Before the txn is committed, another client sends a get
   operation with time T+100. This triggers the txn restart.
3) The client sends another get operation with time T+200. The
   write intent is resolved (and the txn timestamp is pushed to
   T+200), but the actual get operation has not yet been executed (and
   hence the timestamp cache has not been updated).
4) The restarted txn executes the put operation again. The timestamp
   of the operation is T+100, and it will be ignored.

In the above example, a get issued in the same txn after Step 4 will
not see the put.

This commit also change mvccInternalPut to teturn an error if a put
with an earlier epoch came after a put with a newer epoch.